chore(team): split mailbox facade paths
